[DRUP-560] Use the `PrepaidBalanceControllerBase` for developer prepaid balances.
