--- templates/prepaid-balances.html.twig
@@ -1,38 +1,0 @@
-{#
-/**
- * Template for displaying current prepaid balances
- *
- * Available variables:
- * - balances: @todo describe balances var once we switch from direct api call to sdk
- *
- * @see Drupal\apigee_m10n\Controller\BillingController
- */
-#}
-<h3 class="label">{{ 'Current prepaid balance' |t }}</h3>
-<table>
-    <thead>
-        <tr>
-            <th>{{ 'Account Currency' | t }}</th>
-            <th>{{ 'Previous Balance' | t }}</th>
-            <th>{{ 'Credit' | t }}</th>
-            <th>{{ 'Usage' | t }}</th>
-            <th>{{ 'Tax' | t }}</th>
-            <th>{{ 'Current Balance' | t }}</th>
-        </tr>
-    </thead>
-    <tbody>
-        {% if balances %}
-            {% for balance in balances %}
-                {% set currency_code = balance.currency.name %}
-                <tr class="apigee-balance-row-{{ currency_code | lower }}">
-                    <td>{{ balance.currency.name }}</td>
-                    <td>{{ balance.previousBalance | apigee_m10n_format_currency(currency_code) }}</td>
-                    <td>{{ balance.topups | apigee_m10n_format_currency(currency_code) }}</td>
-                    <td>{{ balance.usage | apigee_m10n_format_currency(currency_code) }}</td>
-                    <td>{{ balance.tax | apigee_m10n_format_currency(currency_code) }}</td>
-                    <td>{{ balance.currentBalance | apigee_m10n_format_currency(currency_code) }}</td>
-                </tr>
-            {% endfor %}
-        {% endif %}
-    </tbody>
-</table>
